Abstract away ty/inferty mapping and projecting
