+++ crates/engine/body-ir/src/resolution/infer/type_ref.rs
@@ -8,8 +8,8 @@ use rg_ir_model::items::{
 };
 use rg_text::Name;
 use rg_ty::{
-    GenericArg, RefMutability, Ty,
-    inference::{InferGenericArg, InferNominalTy, InferTy},
+    RefMutability, Ty,
+    inference::{InferGenericArg, InferTy, TypeRefInferenceProjector},
 };
 
 use super::BodyInferenceCtx;
@@ -216,172 +216,15 @@ impl<'subst> InferTypeRefProjector<'subst> {
     /// Resolve a declared type ref shape while preserving substituted inference vars.
     ///
     /// Example: `Option<T>` with `T = ?T` and resolved `Option<unknown>` becomes `Option<?T>`.
-    pub(crate) fn ty_from_type_ref(&self, pattern: &TypeRef, resolved_ty: &Ty) -> InferTy {
-        if let Some(name) = pattern.type_param_name()
-            && let Some(ty) = self.subst.get(name.as_str())
-        {
-            return ty.clone();
-        }
-
-        match (pattern, resolved_ty) {
-            (TypeRef::Unit, Ty::Unit) => InferTy::Unit,
-            (TypeRef::Never, Ty::Never) => InferTy::Never,
-            (TypeRef::Tuple(pattern_fields), Ty::Tuple(resolved_fields))
-                if pattern_fields.len() == resolved_fields.len() =>
-            {
-                InferTy::Tuple(
-                    pattern_fields
-                        .iter()
-                        .zip(resolved_fields)
-                        .map(|(pattern_field, resolved_field)| {
-                            self.ty_from_type_ref(pattern_field, resolved_field)
-                        })
-                        .collect(),
-                )
-            }
-            (
-                TypeRef::Array {
-                    inner: pattern_inner,
-                    len: pattern_len,
-                },
-                Ty::Array {
-                    inner: resolved_inner,
-                    len: resolved_len,
-                },
-            ) if pattern_len == resolved_len => InferTy::Array {
-                inner: Box::new(self.ty_from_type_ref(pattern_inner, resolved_inner)),
-                len: pattern_len.clone(),
-            },
-            (TypeRef::Slice(pattern_inner), Ty::Slice(resolved_inner)) => InferTy::Slice(Box::new(
-                self.ty_from_type_ref(pattern_inner, resolved_inner),
-            )),
-            (
-                TypeRef::Reference {
-                    mutability,
-                    inner: pattern_inner,
-                    ..
-                },
-                Ty::Reference {
-                    mutability: resolved_mutability,
-                    inner: resolved_inner,
-                },
-            ) if InferTypeSubst::ref_mutability(*mutability) == *resolved_mutability => {
-                InferTy::Reference {
-                    mutability: *resolved_mutability,
-                    inner: Box::new(self.ty_from_type_ref(pattern_inner, resolved_inner)),
-                }
-            }
-            (
-                TypeRef::Reference {
-                    mutability,
-                    inner: pattern_inner,
-                    ..
-                },
-                Ty::Unknown,
-            ) => InferTy::Reference {
-                mutability: InferTypeSubst::ref_mutability(*mutability),
-                inner: Box::new(self.ty_from_type_ref(pattern_inner, &Ty::Unknown)),
-            },
-            (TypeRef::Path(path), Ty::Nominal(ty)) => self
-                .nominal_ty_from_path(path, ty.def, &ty.args)
-                .map(InferTy::Nominal)
-                .unwrap_or_else(|| InferTy::from_ty(resolved_ty)),
-            (TypeRef::Path(path), Ty::SelfTy(ty)) => self
-                .nominal_ty_from_path(path, ty.def, &ty.args)
-                .map(InferTy::SelfTy)
-                .unwrap_or_else(|| InferTy::from_ty(resolved_ty)),
-            _ => InferTy::from_ty(resolved_ty),
-        }
+    pub(crate) fn ty_from_type_ref(&mut self, pattern: &TypeRef, resolved_ty: &Ty) -> InferTy {
+        self.project_ty(pattern, resolved_ty)
     }
+}
 
-    /// Project nominal path args, e.g. `Option<T>` into `Option<?T>`.
-    fn nominal_ty_from_path(
-        &self,
-        path: &TypePath,
-        def: rg_ir_model::TypeDefRef,
-        resolved_args: &[GenericArg],
-    ) -> Option<InferNominalTy> {
-        let segment = path.segments.last()?;
-        if segment.args.len() != resolved_args.len() {
-            return None;
-        }
-
-        Some(InferNominalTy {
-            def,
-            args: segment
-                .args
-                .iter()
-                .zip(resolved_args)
-                .map(|(pattern_arg, resolved_arg)| {
-                    self.generic_arg_from_item_arg(pattern_arg, resolved_arg)
-                })
-                .collect(),
-        })
-    }
-
-    /// Project one generic arg from declaration syntax plus its resolved fallback.
-    fn generic_arg_from_item_arg(
-        &self,
-        pattern: &ItemGenericArg,
-        resolved_arg: &GenericArg,
-    ) -> InferGenericArg {
-        match (pattern, resolved_arg) {
-            (ItemGenericArg::Type(pattern_ty), GenericArg::Type(resolved_ty)) => {
-                InferGenericArg::Type(Box::new(self.ty_from_type_ref(pattern_ty, resolved_ty)))
-            }
-            (
-                ItemGenericArg::FnTraitArgs {
-                    params: pattern_params,
-                    ret: pattern_ret,
-                },
-                GenericArg::FnTraitArgs {
-                    params: resolved_params,
-                    ret: resolved_ret,
-                },
-            ) if pattern_params.len() == resolved_params.len() => InferGenericArg::FnTraitArgs {
-                params: pattern_params
-                    .iter()
-                    .zip(resolved_params)
-                    .map(|(pattern_param, resolved_param)| {
-                        self.ty_from_type_ref(pattern_param, resolved_param)
-                    })
-                    .collect(),
-                ret: Box::new(self.ty_from_type_ref(pattern_ret, resolved_ret)),
-            },
-            (
-                ItemGenericArg::AssocType {
-                    name: pattern_name,
-                    ty: Some(pattern_ty),
-                },
-                GenericArg::AssocType {
-                    name: resolved_name,
-                    ty: Some(resolved_ty),
-                },
-            ) if pattern_name == resolved_name => InferGenericArg::AssocType {
-                name: pattern_name.clone(),
-                ty: Some(Box::new(self.ty_from_type_ref(pattern_ty, resolved_ty))),
-            },
-            _ => Self::generic_arg_from_resolved(resolved_arg),
-        }
-    }
-
-    /// Preserve a resolved arg when declaration syntax gives no inference-specific shape.
-    fn generic_arg_from_resolved(resolved_arg: &GenericArg) -> InferGenericArg {
-        match resolved_arg {
-            GenericArg::Type(ty) => InferGenericArg::Type(Box::new(InferTy::from_ty(ty))),
-            GenericArg::Lifetime(lifetime) => InferGenericArg::Lifetime(lifetime.clone()),
-            GenericArg::Const(value) => InferGenericArg::Const(value.clone()),
-            GenericArg::FnTraitArgs { params, ret } => InferGenericArg::FnTraitArgs {
-                params: params.iter().map(InferTy::from_ty).collect(),
-                ret: Box::new(InferTy::from_ty(ret)),
-            },
-            GenericArg::AssocType { name, ty } => InferGenericArg::AssocType {
-                name: name.clone(),
-                ty: ty
-                    .as_ref()
-                    .map(|ty| Box::new(InferTy::from_ty(ty.as_ref()))),
-            },
-            GenericArg::Unsupported(text) => InferGenericArg::Unsupported(text.clone()),
-        }
+impl TypeRefInferenceProjector for InferTypeRefProjector<'_> {
+    /// Substitute declared type params such as `T` with already-bound inference vars.
+    fn replace_written_ty(&mut self, pattern: &TypeRef) -> Option<InferTy> {
+        let name = pattern.type_param_name()?;
+        self.subst.get(name.as_str()).cloned()
     }
 }
