+++ crates/engine/ty/src/inference/mod.rs
@@ -3,10 +3,12 @@
 //! This module keeps solver state separate from the persisted type vocabulary. Callers can carry
 //! shapes such as `Vec<?T>` or `{integer}` during resolution, then finalize them back to `Ty`.
 
+mod family;
 mod instantiate;
 mod model;
 mod table;
 
+pub use family::TypeRefInferenceProjector;
 pub use instantiate::{
     ExplicitTypeArgInstantiationBuilder, GenericReturnInstantiationBuilder,
     UnknownTypeInstantiationBuilder,
